Deploy rewards to Athens2 and 3 (#104)
* Deploy rewards to Athens2

* make public time restrictions
